otelfields: make package private, rename to otelfields
