+++ internal/sinkcores/sentrycore/worker.go
@@ -8,7 +8,7 @@ import (
 	"github.com/cockroachdb/errors"
 	"github.com/getsentry/sentry-go"
 	"github.com/sourcegraph/log/internal/encoders"
-	"github.com/sourcegraph/log/otfields"
+	"github.com/sourcegraph/log/internal/otelfields"
 	"go.uber.org/zap/zapcore"
 )
 
@@ -132,7 +132,7 @@ func (w *worker) capture(errCtx *errorContext) {
 
 	// Extract the service name, if present in the fields.
 	for _, f := range errCtx.Fields {
-		if f.Key == otfields.ResourceFieldKey {
+		if f.Key == otelfields.ResourceFieldKey {
 			if r, ok := f.Interface.(*encoders.ResourceEncoder); ok {
 				tags["resource.service.name"] = r.Name
 				tags["resource.service.version"] = r.Version
